Making the test suite part of the command line program and improving the flag support and help message, combining the scripting into one with better functions. Adding -pedantic, -Wall flags to hunt for C90 incompatibility.
