+++ embroider/geom-circle.dxf
